Add SQL files

+++ src/main/java/org/iish/hsn/invoer/domain/invoer/geb/Geb.java
@@ -5,14 +5,14 @@ import org.iish.hsn.invoer.domain.invoer.Invoer;
 import javax.persistence.*;
 
 @Entity
-@Table(name = "geb")
+@Table(name = "geb") // Is actually not a table, but a view
 public class Geb extends Invoer {
     @Column(name = "GEMNR", nullable = false) private    int    gemnr;
     @Column(name = "GEMNAAM", nullable = false) private  String gemnaam = "";
     @Column(name = "JAAR", nullable = false) private     int    jaar;
     @Column(name = "AKTENR", nullable = false) private   int    aktenr;
     @Column(name = "IDNR", nullable = false) private     int    idnr;
-    @Column(name = "GEBKODE", nullable = false) private  int    gebkode;
+    @Column(name = "GEBKODE", nullable = false) private  long   gebkode; // long, as MySQL turns this into a bigint
     @Column(name = "OVERSAMP", nullable = false) private String oversamp = "";
 
     @Id @Column(name = "RecordID") private Integer recordID;
@@ -37,7 +37,7 @@ public class Geb extends Invoer {
         return idnr;
     }
 
-    public int getGebkode() {
+    public long getGebkode() {
         return gebkode;
     }
